famyoutube/requirements.txt
arrow==0.17.0
asgiref==3.3.1
blessed==1.17.12
certifi==2020.12.5
chardet==4.0.0
croniter==0.3.36
Django==3.1.4
django-mongodb-engine==0.6.0
django-picklefield==3.0.1
django-q==1.3.4
djangorestframework==3.12.2
djangotoolbox==1.8.0
idna==2.10
natsort==7.1.0
pymongo==3.11.2
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
wcwidth==0.2.5
